add support for ESP32 IDF framework.

+++ components/victron_ble/sensor/victron_sensor.cpp
@@ -8,7 +8,7 @@ static const char *const TAG = "victron_ble.sensor";
 
 void VictronSensor::dump_config() {
   LOG_SENSOR("", "Victron Sensor", this);
-  ESP_LOGCONFIG(TAG, "Type %u", this->type_);
+  ESP_LOGCONFIG(TAG, "Type %u", (size_t) this->type_);
 }
 
 void VictronSensor::setup() {
